move migrateImportedFromFrontMatter to sync
