+++ apps/obsidian/src/index.ts
@@ -27,8 +27,6 @@ import { initializeSupabaseSync } from "~/utils/syncDgNodesToSupabase";
 import { FileChangeListener } from "~/utils/fileChangeListener";
 import generateUid from "~/utils/generateUid";
 import { migrateFrontmatterRelationsToRelationsJson } from "~/utils/relationsStore";
-import { collectDiscourseNodesFromVault } from "~/utils/getDiscourseNodes";
-import { spaceUriAndLocalIdToRid } from "~/utils/rid";
 
 export default class DiscourseGraphPlugin extends Plugin {
   settings: Settings = { ...DEFAULT_SETTINGS };
@@ -43,10 +41,6 @@ export default class DiscourseGraphPlugin extends Plugin {
 
     await migrateFrontmatterRelationsToRelationsJson(this).catch((error) => {
       console.error("Failed to migrate frontmatter relations:", error);
-    });
-
-    await this.migrateImportedFromFrontMatter().catch((error) => {
-      console.error("Failed to migrate frontmatter:", error);
     });
 
     if (this.settings.syncModeEnabled === true) {
@@ -399,35 +393,6 @@ export default class DiscourseGraphPlugin extends Plugin {
       }
     });
     this.currentViewActions = [];
-  }
-
-  async migrateImportedFromFrontMatter() {
-    const nodes = await collectDiscourseNodesFromVault(this, true);
-    for (const node of nodes) {
-      if (typeof node.frontmatter.importedFromSpaceUri === "string") {
-        await this.app.fileManager.processFrontMatter(
-          node.file,
-          (frontmatter: Record<string, unknown>) => {
-            const spaceUri = frontmatter.importedFromSpaceUri as string;
-            // note: we fortunately reused the original Id here.
-            const nodeId = frontmatter.nodeInstanceId;
-            if (typeof nodeId !== "string") {
-              console.error(
-                `error: missing nodeInstanceId on node ${node.file.path}`,
-              );
-              return;
-            }
-            try {
-              const rid = spaceUriAndLocalIdToRid(spaceUri, nodeId, "note");
-              frontmatter.importedFromRid = rid;
-              delete frontmatter.importedFromSpaceUri;
-            } catch (error) {
-              console.error(error);
-            }
-          },
-        );
-      }
-    }
   }
 
   async onunload() {
